requirements.txt
numpy
pandas
google-generativeai
streamlit
watchdog